fixes in FileDownload
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@1004 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/common/FileDownload.cpp
@@ -177,7 +177,7 @@ void CFileDownload::ProcessDownload()
 		Lock();
 		if (tHttp.GetError().iError == HTTP_FILE_NOT_FOUND)  {
 			iCurrentServer++;
-			if (iCurrentServer < tDownloadServers->size())  {
+			if ((size_t)iCurrentServer < tDownloadServers->size())  {
 				tHttp.RequestData((*tDownloadServers)[iCurrentServer] + sFileName);  // Request the file
 				iState = FILEDL_INITIALIZING;
 				Unlock();
